fix: make import replacement support `import type`
We now need to support `import type` and multiple import lines, such
such that there can be an `import type { ...} from '...intervene'` and
a second `import { ... } from 'intervene'` line. This also meant we
could no longer rely on importing `ProxyConfig` as the confirmation
that this was indeed an intervene import line. So we now rely on there
only being intervene known imports in the statement _and_ that the word
`intervene` is somewhere in the import path.

Note that this is related to the TypeScript update, as TS4 introduces
`import type`.