tests: fix end-to-end tests to run against vercel
Since the name change, we needed to update the hostname to the new
vercel.app hostname for the tests, from
github.com/bruderstein/intervene-test.

Note that this still doesn't work against production, as the
intervene-test project needs updating to run with the new vercel
environment (was previous now.sh, which was renamed).

But running the tests against that repo running locally works now